pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.1.0",
    "langchain-deepseek>=1.0.1",
    "openai>=2.8.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
